Refactored file structure for chat.
