+++ app/javascript/src/views/Chat/alert.jsx
