+++ app/javascript/src/views/Chat/article.jsx
