+++ app/javascript/src/views/Chat/channelDetails.jsx
