+++ app/javascript/src/views/Chat/content.jsx
