+++ app/javascript/src/views/Chat/messages/errorMessage.jsx
