+++ app/javascript/src/views/Chat/userDetails.jsx
