+++ app/javascript/src/views/Chat/util.js
