+++ app/javascript/src/views/Chat/view.jsx
